chore: move start deletion process dialog and fix mocha arrows
