+++ AdminApi/src/AdminApi/ClientApp/src/app/app.module.ts
@@ -45,9 +45,9 @@ import { PageNotFoundComponent } from "./components/error/page-not-found/page-no
 import { DeletionProcessesComponent } from "./components/identity/identity-details/deletion-processes/deletion-processes.component";
 import { CancelDeletionProcessDialogComponent } from "./components/identity/identity-details/deletion-processes/dp-details/cancel-dp-dialog/cancel-dp-dialog.component";
 import { DeletionProcessDetailsComponent } from "./components/identity/identity-details/deletion-processes/dp-details/dp-details.component";
+import { StartDeletionProcessDialogComponent } from "./components/identity/identity-details/deletion-processes/start-deletion-process-dialog/start-deletion-process-dialog.component";
 import { IdentityDetailsRelationshipsComponent } from "./components/identity/identity-details/identity-details-relationships/identity-details-relationships.component";
 import { IdentityDetailsComponent } from "./components/identity/identity-details/identity-details.component";
-import { StartDeletionProcessDialogComponent } from "./components/identity/identity-details/start-deletion-process-dialog/start-deletion-process-dialog.component";
 import { IdentityListComponent } from "./components/identity/identity-list/identity-list.component";
 import { AssignQuotasDialogComponent } from "./components/quotas/assign-quotas-dialog/assign-quotas-dialog.component";
 import { CreateTierDialogComponent } from "./components/quotas/tier/create-tier-dialog/create-tier-dialog.component";
